+++ cmd/scenarigo/cmd/config/init_test.go
@@ -24,10 +24,19 @@ func TestInitRun(t *testing.T) {
 	})
 	t.Run("failure", func(t *testing.T) {
 		t.Run("already exist", func(t *testing.T) {
+			wd, err := os.Getwd()
+			if err != nil {
+				t.Fatal(err)
+			}
 			dir := t.TempDir()
 			if err := os.Chdir(dir); err != nil {
 				t.Fatal(err)
 			}
+			defer func() {
+				if err := os.Chdir(wd); err != nil {
+					t.Fatal(err)
+				}
+			}()
 			if err := initRun(nil, []string{}); err != nil {
 				t.Fatalf("failed to init: %s", err)
 			}
