+++ cmd/scenarigo/cmd/run_test.go
@@ -113,58 +113,3 @@ ok  	scenarios/pass.yaml	0.000s
 		})
 	}
 }
-
-func TestLoadConfig(t *testing.T) {
-	tests := map[string]struct {
-		filename string
-		cd       string
-		found    bool
-		fail     bool
-	}{
-		"default (not found)": {},
-		"default (found)": {
-			cd:    "./testdata",
-			found: true,
-		},
-		"specify file": {
-			filename: "./testdata/scenarigo.yaml",
-			found:    true,
-		},
-		"specify file (not found)": {
-			filename: "./testdata/.invalid.yaml",
-			fail:     true,
-		},
-	}
-	for name, test := range tests {
-		test := test
-		t.Run(name, func(t *testing.T) {
-			if test.cd != "" {
-				wd, err := os.Getwd()
-				if err != nil {
-					t.Fatal(err)
-				}
-				if err := os.Chdir(test.cd); err != nil {
-					t.Fatal(err)
-				}
-				defer func() {
-					if err := os.Chdir(wd); err != nil {
-						t.Fatal(err)
-					}
-				}()
-			}
-			cfg, err := loadConfig(test.filename)
-			if test.fail && err == nil {
-				t.Fatal("no error")
-			}
-			if !test.fail && err != nil {
-				t.Fatalf("unexpected error: %s", err)
-			}
-			if test.found && cfg == nil {
-				t.Error("config not found")
-			}
-			if !test.found && cfg != nil {
-				t.Error("unknown config")
-			}
-		})
-	}
-}
